it doesn't work :(

+++ final/FinalProject/Assignment.cs
@@ -2,8 +2,8 @@ public abstract class Assignment
 {
     private string _name;
     private DateTime _due;
-    private int _points;
-    public Assignment(string name, DateTime due_date, int max_points)
+    private double _points;
+    public Assignment(string name, DateTime due_date, double max_points)
     {
         _name = name;
         _due = due_date;
